py-requirements.txt
flask
imutils
paho-mqtt
pyaudio
pyttsx3
uwsgi
scikit-learn
face_recognition